requirements.txt
pandas==2.2.2
python-dateutil==2.9.0.post0
typer==0.12.3
rich==13.9.1
tqdm==4.66.4
mail-parser==3.15.0
spacy==3.7.5
python-dotenv==1.0.1
pymysql==1.1.1
streamlit==1.36.0
